Separated boolean optimizers from bitwise unary operator optimizers. Renamed OptCmp6 to OptBoolCmp.
