+++ src/cc65/coptcmp.c
@@ -44,127 +44,8 @@
 
 
 /*****************************************************************************/
-/*                                   Data                                    */
-/*****************************************************************************/
-
-
-
-/* Table used to invert a condition, indexed by condition */
-static const unsigned char CmpInvertTab [] = {
-    CMP_NE, CMP_EQ,
-    CMP_LE, CMP_LT, CMP_GE, CMP_GT,
-    CMP_ULE, CMP_ULT, CMP_UGE, CMP_UGT
-};
-
-
-
-/*****************************************************************************/
 /*                             Helper functions                              */
 /*****************************************************************************/
-
-
-
-static void ReplaceCmp (CodeSeg* S, unsigned I, cmp_t Cond)
-/* Helper function for the replacement of routines that return a boolean
-** followed by a conditional jump. Instead of the boolean value, the condition
-** codes are evaluated directly.
-** I is the index of the conditional branch, the sequence is already checked
-** to be correct.
-*/
-{
-    CodeEntry* N;
-    CodeLabel* L;
-
-    /* Get the entry */
-    CodeEntry* E = CS_GetEntry (S, I);
-
-    /* Replace the conditional branch */
-    switch (Cond) {
-
-        case CMP_EQ:
-            CE_ReplaceOPC (E, OP65_JEQ);
-            break;
-
-        case CMP_NE:
-            CE_ReplaceOPC (E, OP65_JNE);
-            break;
-
-        case CMP_GT:
-            /* Replace by
-            **     beq @L
-            **     jpl Target
-            ** @L: ...
-            */
-            if ((N = CS_GetNextEntry (S, I)) == 0) {
-                /* No such entry */
-                Internal ("Invalid program flow");
-            }
-            L = CS_GenLabel (S, N);
-            N = NewCodeEntry (OP65_BEQ, AM65_BRA, L->Name, L, E->LI);
-            CS_InsertEntry (S, N, I);
-            CE_ReplaceOPC (E, OP65_JPL);
-            break;
-
-        case CMP_GE:
-            CE_ReplaceOPC (E, OP65_JPL);
-            break;
-
-        case CMP_LT:
-            CE_ReplaceOPC (E, OP65_JMI);
-            break;
-
-        case CMP_LE:
-            /* Replace by
-            **     jmi Target
-            **     jeq Target
-            */
-            CE_ReplaceOPC (E, OP65_JMI);
-            L = E->JumpTo;
-            N = NewCodeEntry (OP65_JEQ, AM65_BRA, L->Name, L, E->LI);
-            CS_InsertEntry (S, N, I+1);
-            break;
-
-        case CMP_UGT:
-            /* Replace by
-            **     beq @L
-            **     jcs Target
-            ** @L: ...
-            */
-            if ((N = CS_GetNextEntry (S, I)) == 0) {
-                /* No such entry */
-                Internal ("Invalid program flow");
-            }
-            L = CS_GenLabel (S, N);
-            N = NewCodeEntry (OP65_BEQ, AM65_BRA, L->Name, L, E->LI);
-            CS_InsertEntry (S, N, I);
-            CE_ReplaceOPC (E, OP65_JCS);
-            break;
-
-        case CMP_UGE:
-            CE_ReplaceOPC (E, OP65_JCS);
-            break;
-
-        case CMP_ULT:
-            CE_ReplaceOPC (E, OP65_JCC);
-            break;
-
-        case CMP_ULE:
-            /* Replace by
-            **     jcc Target
-            **     jeq Target
-            */
-            CE_ReplaceOPC (E, OP65_JCC);
-            L = E->JumpTo;
-            N = NewCodeEntry (OP65_JEQ, AM65_BRA, L->Name, L, E->LI);
-            CS_InsertEntry (S, N, I+1);
-            break;
-
-        default:
-            Internal ("Unknown jump condition: %d", Cond);
-
-    }
-
-}
 
 
 
@@ -201,68 +82,6 @@ static int GetCmpRegVal (const CodeEntry* E)
         default:       Internal ("Invalid opcode in GetCmpRegVal");
                        return 0;  /* Not reached */
     }
-}
-
-
-
-/*****************************************************************************/
-/*             Remove calls to the bool transformer subroutines              */
-/*****************************************************************************/
-
-
-
-unsigned OptBoolTrans (CodeSeg* S)
-/* Try to remove the call to boolean transformer routines where the call is
-** not really needed.
-*/
-{
-    unsigned Changes = 0;
-
-    /* Walk over the entries */
-    unsigned I = 0;
-    while (I < CS_GetEntryCount (S)) {
-
-        CodeEntry* N;
-        cmp_t Cond;
-
-        /* Get next entry */
-        CodeEntry* E = CS_GetEntry (S, I);
-
-        /* Check for a boolean transformer */
-        if (E->OPC == OP65_JSR                           &&
-            (Cond = FindBoolCmpCond (E->Arg)) != CMP_INV &&
-            (N = CS_GetNextEntry (S, I)) != 0            &&
-            (N->Info & OF_ZBRA) != 0) {
-
-            /* Make the boolean transformer unnecessary by changing the
-            ** the conditional jump to evaluate the condition flags that
-            ** are set after the compare directly. Note: jeq jumps if
-            ** the condition is not met, jne jumps if the condition is met.
-            ** Invert the code if we jump on condition not met.
-            */
-            if (GetBranchCond (N->OPC) == BC_EQ) {
-                /* Jumps if condition false, invert condition */
-                Cond = CmpInvertTab [Cond];
-            }
-
-            /* Check if we can replace the code by something better */
-            ReplaceCmp (S, I+1, Cond);
-
-            /* Remove the call to the bool transformer */
-            CS_DelEntry (S, I);
-
-            /* Remember, we had changes */
-            ++Changes;
-
-        }
-
-        /* Next entry */
-        ++I;
-
-    }
-
-    /* Return the number of changes made */
-    return Changes;
 }
 
 
@@ -671,68 +490,6 @@ unsigned OptCmp5 (CodeSeg* S)
             }
 
             ++Changes;
-        }
-
-        /* Next entry */
-        ++I;
-
-    }
-
-    /* Return the number of changes made */
-    return Changes;
-}
-
-
-
-unsigned OptCmp6 (CodeSeg* S)
-/* Search for calls to compare subroutines followed by a conditional branch
-** and replace them by cheaper versions, since the branch means that the
-** boolean value returned by these routines is not needed (we may also check
-** that explicitly, but for the current code generator it is always true).
-*/
-{
-    unsigned Changes = 0;
-
-    /* Walk over the entries */
-    unsigned I = 0;
-    while (I < CS_GetEntryCount (S)) {
-
-        CodeEntry* N;
-        cmp_t Cond;
-
-        /* Get next entry */
-        CodeEntry* E = CS_GetEntry (S, I);
-
-        /* Check for the sequence */
-        if (E->OPC == OP65_JSR                          &&
-            (Cond = FindTosCmpCond (E->Arg)) != CMP_INV &&
-            (N = CS_GetNextEntry (S, I)) != 0           &&
-            (N->Info & OF_ZBRA) != 0                    &&
-            !CE_HasLabel (N)) {
-
-            /* The tos... functions will return a boolean value in a/x and
-            ** the Z flag says if this value is zero or not. We will call
-            ** a cheaper subroutine instead, one that does not return a
-            ** boolean value but only valid flags. Note: jeq jumps if
-            ** the condition is not met, jne jumps if the condition is met.
-            ** Invert the code if we jump on condition not met.
-            */
-            if (GetBranchCond (N->OPC) == BC_EQ) {
-                /* Jumps if condition false, invert condition */
-                Cond = CmpInvertTab [Cond];
-            }
-
-            /* Replace the subroutine call. */
-            E = NewCodeEntry (OP65_JSR, AM65_ABS, "tosicmp", 0, E->LI);
-            CS_InsertEntry (S, E, I+1);
-            CS_DelEntry (S, I);
-
-            /* Replace the conditional branch */
-            ReplaceCmp (S, I+1, Cond);
-
-            /* Remember, we had changes */
-            ++Changes;
-
         }
 
         /* Next entry */
